more changes, added a server (express / nodejs)
